pyproject.toml
[project]
name = "goldfish"
version = "0.1.0"
description = "MCP server for ML experimentation - solves the goldfish memory problem for AI agents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=0.1.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
goldfish = "goldfish.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/goldfish"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
